requirements.txt
streamlit
pandas
numpy
google-api-python-client
google-auth
google-auth-oauthlib
plotly
python-docx
kaleido